+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/provider/BookmarksStorageSuggestionProvider.kt
@@ -6,6 +6,7 @@ package mozilla.components.feature.awesomebar.provider
 
 import android.graphics.drawable.Drawable
 import androidx.annotation.VisibleForTesting
+import androidx.core.net.toUri
 import mozilla.components.browser.icons.BrowserIcons
 import mozilla.components.browser.icons.IconRequest
 import mozilla.components.concept.awesomebar.AwesomeBar
@@ -17,7 +18,17 @@ import mozilla.components.feature.awesomebar.facts.emitBookmarkSuggestionClicked
 import mozilla.components.feature.session.SessionUseCases
 import java.util.UUID
 
-private const val BOOKMARKS_SUGGESTION_LIMIT = 20
+/**
+ * Return up to 20 bookmarks suggestions by default.
+ */
+@VisibleForTesting
+internal const val BOOKMARKS_SUGGESTION_LIMIT = 20
+
+/**
+ * Default suggestions limit multiplier when needing to filter results by an external url filter.
+ */
+@VisibleForTesting
+internal const val BOOKMARKS_RESULTS_TO_FILTER_SCALE_FACTOR = 10
 
 /**
  * A [AwesomeBar.SuggestionProvider] implementation that provides suggestions based on the bookmarks
@@ -33,6 +44,7 @@ private const val BOOKMARKS_SUGGESTION_LIMIT = 20
  * highest scored suggestion URL.
  * @param showEditSuggestion optional parameter to specify if the suggestion should show the edit button
  * @param suggestionsHeader optional parameter to specify if the suggestion should have a header
+ * @param resultsHostFilter Optional filter for the host url of the suggestions to show.
  */
 class BookmarksStorageSuggestionProvider(
     @get:VisibleForTesting internal val bookmarksStorage: BookmarksStorage,
@@ -42,6 +54,7 @@ class BookmarksStorageSuggestionProvider(
     private val engine: Engine? = null,
     private val showEditSuggestion: Boolean = true,
     private val suggestionsHeader: String? = null,
+    @get:VisibleForTesting val resultsHostFilter: String? = null,
 ) : AwesomeBar.SuggestionProvider {
     override val id: String = UUID.randomUUID().toString()
 
@@ -56,15 +69,42 @@ class BookmarksStorageSuggestionProvider(
             return emptyList()
         }
 
-        val suggestions = bookmarksStorage.searchBookmarks(text, BOOKMARKS_SUGGESTION_LIMIT)
-            .filter { it.url != null }
-            .distinctBy { it.url }
-            .sortedBy { it.guid }
+        val suggestions = when (resultsHostFilter) {
+            null -> getBookmarksSuggestions(text)
+            else -> getBookmarksSuggestionsFromHost(resultsHostFilter, text)
+        }
 
         suggestions.firstOrNull()?.url?.let { url -> engine?.speculativeConnect(url) }
 
         return suggestions.into()
     }
+
+    /**
+     * Get up to [BOOKMARKS_SUGGESTION_LIMIT] bookmarks matching [query].
+     *
+     * @param query String to filter bookmarks' title or URL by.
+     */
+    private suspend fun getBookmarksSuggestions(query: String) = bookmarksStorage
+        .searchBookmarks(query, BOOKMARKS_SUGGESTION_LIMIT)
+        .filter { it.url != null }
+        .distinctBy { it.url }
+        .sortedBy { it.guid }
+
+    /**
+     * Get up to [BOOKMARKS_SUGGESTION_LIMIT] bookmarks matching [query] from the indicated [host].
+     *
+     * @param query String to filter bookmarks' title or URL by.
+     * @param host URL host to filter all bookmarks' URL host by.
+     */
+    private suspend fun getBookmarksSuggestionsFromHost(host: String, query: String) = bookmarksStorage
+        .searchBookmarks(host, BOOKMARKS_SUGGESTION_LIMIT * BOOKMARKS_RESULTS_TO_FILTER_SCALE_FACTOR)
+        .filter {
+            it.url?.toUri()?.host?.equals(host) ?: true &&
+                (it.url?.contains(query, true) ?: false || it.title?.contains(query, true) ?: false)
+        }
+        .distinctBy { it.url }
+        .sortedBy { it.guid }
+        .take(BOOKMARKS_SUGGESTION_LIMIT)
 
     /**
      * Expects list of BookmarkNode to be specifically of bookmarks (e.g. nodes with a url).