+++ mobile/android/android-components/components/feature/awesomebar/src/main/java/mozilla/components/feature/awesomebar/provider/HistoryStorageSuggestionProvider.kt
@@ -5,6 +5,7 @@
 package mozilla.components.feature.awesomebar.provider
 
 import androidx.annotation.VisibleForTesting
+import androidx.core.net.toUri
 import mozilla.components.browser.icons.BrowserIcons
 import mozilla.components.browser.icons.IconRequest
 import mozilla.components.concept.awesomebar.AwesomeBar
@@ -19,6 +20,12 @@ import java.util.UUID
  * Return 20 history suggestions by default.
  */
 const val DEFAULT_HISTORY_SUGGESTION_LIMIT = 20
+
+/**
+ * Default suggestions limit multiplier when needing to filter results by an external url filter.
+ */
+@VisibleForTesting
+internal const val HISTORY_RESULTS_TO_FILTER_SCALE_FACTOR = 10
 
 /**
  * A [AwesomeBar.SuggestionProvider] implementation that provides suggestions based on the browsing
@@ -36,6 +43,7 @@ const val DEFAULT_HISTORY_SUGGESTION_LIMIT = 20
  * defaults to [DEFAULT_HISTORY_SUGGESTION_LIMIT]
  * @param showEditSuggestion optional parameter to specify if the suggestion should show the edit button
  * @param suggestionsHeader optional parameter to specify if the suggestion should have a header
+ * @param resultsHostFilter Optional filter for the host url of the suggestions to show.
  */
 class HistoryStorageSuggestionProvider(
     @get:VisibleForTesting internal val historyStorage: HistoryStorage,
@@ -45,6 +53,7 @@ class HistoryStorageSuggestionProvider(
     @get:VisibleForTesting internal var maxNumberOfSuggestions: Int = DEFAULT_HISTORY_SUGGESTION_LIMIT,
     private val showEditSuggestion: Boolean = true,
     private val suggestionsHeader: String? = null,
+    @get:VisibleForTesting val resultsHostFilter: String? = null,
 ) : AwesomeBar.SuggestionProvider {
 
     override val id: String = UUID.randomUUID().toString()
@@ -60,12 +69,10 @@ class HistoryStorageSuggestionProvider(
             return emptyList()
         }
 
-        // In case of duplicates we want to pick the suggestion with the highest score.
-        // See: https://github.com/mozilla/application-services/issues/970
-        val suggestions = historyStorage.getSuggestions(text, maxNumberOfSuggestions)
-            .sortedByDescending { it.score }
-            .distinctBy { it.id }
-            .take(maxNumberOfSuggestions)
+        val suggestions = when (resultsHostFilter) {
+            null -> getHistorySuggestions(text)
+            else -> getHistorySuggestionsFromHost(resultsHostFilter, text)
+        }
 
         suggestions.firstOrNull()?.url?.let { url -> engine?.speculativeConnect(url) }
 
@@ -84,11 +91,46 @@ class HistoryStorageSuggestionProvider(
     }
 
     /**
+     * Get the maximum number of suggestions that will be provided.
+     */
+    @VisibleForTesting
+    fun getMaxNumberOfSuggestions() = maxNumberOfSuggestions
+
+    /**
      * Reset maximum number of suggestions to default.
      */
     fun resetToDefaultMaxSuggestions() {
         maxNumberOfSuggestions = DEFAULT_HISTORY_SUGGESTION_LIMIT
     }
+
+    /**
+     * Get up to [maxNumberOfSuggestions] history suggestions matching [query]].
+     *
+     * @param query String to filter history entry's title or URL by.
+     */
+    private fun getHistorySuggestions(query: String) = historyStorage
+        .getSuggestions(query, maxNumberOfSuggestions)
+        // In case of duplicates we want to pick the suggestion with the highest score.
+        // See: https://github.com/mozilla/application-services/issues/970
+        .sortedByDescending { it.score }
+        .distinctBy { it.id }
+        .take(maxNumberOfSuggestions)
+
+    /**
+     * Get up to [maxNumberOfSuggestions] history suggestions matching [query] from the indicated [host].
+     *
+     * @param query String to filter history entry's title or URL by.
+     * @param host URL host to filter all history entry's URL host by.
+     */
+    private fun getHistorySuggestionsFromHost(host: String, query: String) = historyStorage
+        .getSuggestions(host, maxNumberOfSuggestions * HISTORY_RESULTS_TO_FILTER_SCALE_FACTOR)
+        .sortedByDescending { it.score }
+        .distinctBy { it.id }
+        .filter {
+            it.url.toUri().host == host &&
+                (it.url.contains(query, true) || it.title?.contains(query, true) ?: false)
+        }
+        .take(maxNumberOfSuggestions)
 }
 
 internal suspend fun Iterable<SearchResult>.into(